+++ service/controller/resource/drainerinitializer/delete.go
@@ -1,4 +1,4 @@
-package drainer
+package drainerinitializer
 
 import (
 	"context"