+++ CHANGELOG.md
@@ -11,11 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add Control Plane drainer controller.
 - Add VPC ID to AWSCluster CR status.
 
 ### Fixed
 
 - Fix aws operator policy for latest node pools version.
+- Make encryption key lookup graceful during cluster creation.
+
 
 
 
@@ -35,6 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use architect orb.
 - Fix subnet allocation for Availability Zones.
 - Switch to AWS CNI
+
+
 
 ## [8.2.1] - 2020-03-20